refactor: add `Hooks` folder
